requirements.txt
# Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1

# Image Processing
Pillow==11.3.0

# Configuration
python-dotenv==1.0.0

# For health checks in Docker
gunicorn==21.2.0
